giveup apollo method and test finish
